Crate sonr

Simple Oppinionated Networking in Rust

Sonr is built on top of mio, with the goal of making networking in Rust a bit easier.

Modules

connections
Connections and Sessions

errors
SONR default Error

server
Servers and listeners

Structs

Event
An readiness event returned by Poll::poll.

Events
A collection of readiness events.

Poll
Polls for readiness events on all registered values.

PreVec
PreVec: a collection that allows inserts at a specific index with an optional offsets. A PreVec can be created with either capacity or capacity_with_offset.

Ready
A set of readiness event kinds

Token
Associates readiness notifications with Evented handles.

Traits

Evented
A value that may be registered with Poll