Struct sonr::connections::Session

pub struct Session<C> { /* fields omitted */ }

Unlike a connection the session has an internal buffer.

A session is required to be registered with a poll handle or it won't listen to events.

let poll = Poll::new()?;
let mut events = Events::with_capacity(1024);
let connection = TcpConnection::connect("127.0.0.1", 5000)?;
let mut session = Session::new(connection);
let token = Token(1);
 
// The session will listen to **read** events
session.register_readable(&poll, token)?;

// The session will listen to **write** events as well
session.reregister_writable(&poll, token)?;

Methods

impl<C: Connection> Session<C>

pub fn new(connection: C) -> Self

Create a new session from a connection

pub fn with_capacity(connection: C, capacity: usize) -> Self

Create a new session with the capacity for the internal buffer

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn slice(&self) -> &[u8]

Expose a slice of the internal buffer with data from the last fill_buffer call. The buffer will be empty until fill_buffer is called

pub fn write_buf(
&mut self,
poll: &Poll,
token: Token,
buf: &[u8]
) -> EventResult

Write data into a slice. Call this on an event with event.readiness().is_writable() or the call will block If the Session is handled outside of Sessions then make sure that:

The token is the same token as the session was registered with.

The session doesn't try to read / write again in the case of a WriteResult::ConnectionRemoved

When using a Session together with the Sessions collection, the collection will manage the removal of the session.

pub fn read_buf(&mut self, poll: &Poll, token: Token) -> EventResult

Read data into the internal buffer. Call this on an event with event.readiness().is_readable(). Calling this when the Session is not registered as readable or the token doesn't match the token the Session was registered with, will block.

pub fn register_writable(&mut self, poll: &Poll, token: Token) -> Result<()>

Register the underlying connection as writable. Either register_readable or register_writable should be called, and only once.

If the Session is added to a Sessions collection then all registration calls should be managed through Sessions and never directly on the Session instance.

See reregister_readable and reregister_writable on Sessions

pub fn register_readable(&mut self, poll: &Poll, token: Token) -> Result<()>

Register the underlying connection as readable. See register_writable as the same information applies

pub fn reregister_writable(&mut self, poll: &Poll, token: Token) -> Result<()>

Reregister the underlying connection as writable. Unlike register_readable and register_writable, reregister can be called multiple times.

In fact once an event is exhausted either [reregister_readable] and / or [reregister_writable] should be called as the Session will no longer listen to an event once it's complete.

pub fn reregister_readable(&mut self, poll: &Poll, token: Token) -> Result<()>

Reregister the underlying connection as readable. See register_writable as same information applies, excpet for reading rather than writing to the underlying connection.