Struct sonr::connections::Sessions

pub struct Sessions<C> { /* fields omitted */ }

Sessions handles multiple connections.

let poll = Poll::new()?;
let mut events = Events::with_capacity(1024);
let mut sessions = Sessions::<TcpConnection>::with_capacity(1_000_000);
 
// Set the server token to be outside the session capacity
// to ensure the servers token is never the same as a `sessions`
let server_token = 1_000_000 + 1;

// Create a server instance...
let mut tcp_server = Server::new(tcp_listener("127.0.0.1", 5000)?, server_token, &poll);
// ... and listen for incoming connections
tcp_server.listen()?;

// listen for incoming events.
// this would normally happen in a `loop { ... }`
let timeout = Duration::from_millis(1); // instantly time out
poll.poll(&mut events, Some(timeout));

for event in &events {
    if event.token() == tcp_server.token() {
        let stream = tcp_server.accept()?;
        let session = Session::new(stream.into());
        sessions.add(&poll, session);
    } else {
        if event.readiness().is_readable() {
            sessions.try_read(&poll, &event)
            .and_then(|buf| {
                // `buf` contains data read from the 
                // underlying connection.
                // this will be called until the
                // connection would block (or fails)
            })
            .done(|byte_count| {
                // Done is called once and only once if
                // anything was read, with the total number
                // of bytes read

                // reregister the connection
                // to listen to `write` events
                sessions.reregister_writable(
                    &poll,
                    event.token()
                );
            });
        }

        if event.readiness().is_writable() {
            let buf = b"payload";
            sessions.try_write(&poll, &event)
                .with(|| { &buf[..] }) // payload to write
                .done(|byte_count| { println!("wrote {} bytes", byte_count); });
        }
    }
}

// The session will listen to **read** events

Methods

impl<C> Sessions<C> where
C: Connection,

pub fn with_capacity(capacity: usize) -> Self

Create a Session with a max number of connections. Sessions are built on top of a PreVec which prevents growth by default. This is to ensure that two Sessions don't overlap.

pub fn with_capacity_and_offset(capacity: usize, offset: usize) -> Self

Create a Session with a max number of connections and a connection offset. This is useful in the event of multiple Sessions objects being registered with the same Poll handle.

let max_con = 100_000;

// Adding a `Session` to tcp_sessions would give it 
// a token of 1
let mut tcp_sessions = Sessions::<TcpConnection>::with_capacity(max_con);

// Adding a `Session` to uds_sessions would give it
// a token of 100_000
let mut uds_sessions = Sessions::<UdsConnection>::with_capacity_and_offset(100, max_con);

pub fn add(
&mut self,
poll: &Poll,
session: Session<C>
) -> Result<(Token, &mut Session<C>)>

Add a Session and register it with the poll handle. The Session is neither listening for read or write events at this point.

To receive any events for the session call either register_readable and / or register_writable.

pub fn remove<T: Into<usize>>(&mut self, token: T)

Remove and subsequently close the Session and underlying connection. It is safe to call remove multiple times.

pub fn get_mut<T: Into<usize>>(&mut self, token: T) -> Option<&mut Session<C>>

Get a mutable reference to a Session

pub fn reregister_writable(&mut self, poll: &Poll, token: Token) -> Result<()>

Reregister the underlying connection as writable. Calling reregister_writable will reregister interest for write events with the Poll handle.

Writing should not happen outside of a write event

pub fn reregister_readable(&mut self, poll: &Poll, token: Token) -> Result<()>

Register the underlying connection as readable. Calling reregister_readable will reregister interest for read events with the Poll handle.

Reading outside of a read event will block

pub fn try_read<'session>(
&'session mut self,
poll: &'session Poll,
event: &'session Event
) -> EventReader<'session, C>

Try to read from the underlying connection matching the token of the event. It's safe to call try_read without checking whether the event is readable or not as this will be checked before reading. In the event of calling try_read on an event that is not readable, the EventReader will return ReadResult::NoEvent.

try_read will return an EventReader. The EventReader will continuosly call and_then with a slice (&[u8]) for as long as the connection is producing data.

The slice needs to be handled in each and_then call as it won't be preserved between calls.

For more information about how to handle read events see EventReader.

pub fn try_write<'session>(
&'session mut self,
poll: &'session Poll,
event: &'session Event
) -> EventWriter<'session, C>

Try to write to the underlying connection matching the token of the event. Like try_read it's safe to call without checking if the event is writable.

try_write will return an EventWriter. The EventWriter will call with accepting a closure wich should return the buffer to write.

For more information about how to handle write events see EventWriter.