Module sonr::reactor

Reactors are the heart of Sonr and work by pushing data down a chain of reactors.

 use std::io::Write;
 use std::thread;
 use sonr::prelude::*;
 use sonr::errors::Result;
 use sonr::sync::queue::{ReactiveQueue, ReactiveDeque};
 use sonr::net::tcp::{ReactiveTcpListener, TcpStream, ReactiveTcpStream};

 // -----------------------------------------------------------------------------
 // 		- Writer -
 // 		Writer "bye" and drop the connection
 // -----------------------------------------------------------------------------
 struct Writer;

 impl Reactor for Writer {
     type Input = TcpStream;
     type Output = ();

     fn react(&mut self, reaction: Reaction<Self::Input>) -> Reaction<Self::Output> {
         use Reaction::*;
         match reaction {
             Value(mut stream) => {
                 stream.write(b"bye\n");
                 Continue
             }
             Event(ev) => Continue,
             Continue => Continue,
         }
     }
 }

 fn main() -> Result<()> {
     System::init()?;

     // Listen for incoming connections
     let listener = ReactiveTcpListener::bind("127.0.0.1:8000")?.map(|(s, _)| s);
     // Connection queue for connections to be sent to another thread.
     let mut queue = ReactiveQueue::unbounded();

     for _ in 0..4 {
         let deque = queue.deque();
         thread::spawn(move || -> Result<()> {
             System::init()?;
             let deque = ReactiveDeque::new(deque)?;
             let writer = Writer;
             let run = deque.chain(writer);
             System::start(run)?;
             Ok(())
         });
     }

     let run = listener.chain(queue);
     System::start(run)?;
     Ok(())
 }

Modules

consumers
Consumers consume the output of other reactors.

producers
There are two producers:

Structs

And
Use and to run more than one (evented) reactor in parallel.

Chain
Chain two Reactors together, making the output of the first reactor the input of the second.

EventedReactor
The EventedReactor is driven by the System.

Map
Map will capture the Reaction::Value(val) returned by react and apply the provided closure on the value.

Or

Enums

Either
Either A or B Used as output when a_reactor.or(another_reactor)

Reaction
Input / Output of a Reactor.

Traits

Reactor
A reactor reacts to a Reaction and returns a Reaction.