Struct sonr::reactor::Map

pub struct Map<S, F, T> { /* fields omitted */ }

Map will capture the Reaction::Value(val) returned by react and apply the provided closure on the value.

The returned value from the map is the new Output of the Reactor.

 use sonr::net::tcp::{ReactiveTcpListener, TcpStream};

 fn main() -> Result<()> {
     let system_signals = System::init()?;

     let listener = ReactiveTcpListener::bind("127.0.0.1:5557")?;
     let run = listener.map(|(strm, addr)| {
         // Return the tcp stream, ignoring the SocketAddr
         strm
     });

     System::start(run)?;
     Ok(())
 }

Methods

impl<S, F, T> Map<S, F, T>

pub fn new(source: S, callback: F) -> Self

Create a new map from a reactor and a closure.