Trait sonr::reactor::Reactor

pub trait Reactor: Sized {
    type Output;
    type Input;
    fn react(
        &mut self, 
        reaction: Reaction<Self::Input>
    ) -> Reaction<Self::Output> { ... }

    fn chain<T: Reactor>(self, to: T) -> Chain<Self, T> { ... }

    fn and<C: Reactor>(self, second: C) -> And<Self, C> { ... }

    fn map<F, T>(self, callback: F) -> Map<Self, F, T> { ... }
}

A reactor ...

Associated Types

type Output

The output passed to the next reactor in the chain.

type Input

Expected input type from the previous reactor in the chain.

Loading content...
Provided methods

fn react(&mut self, reaction: Reaction<Self::Input>) -> Reaction<Self::Output>

The generated output is passed as the input to the next reactor in the chain.

react is called repeatedly until the reaction returns Reaction::Continue

fn chain<T: Reactor>(self, to: T) -> Chain<Self, T>

Chain two reactors together. The output of the first reactor is the input of the second reactor.

fn and<C: Reactor>(self, second: C) -> And<Self, C>

Run two reactors independent of each other.

use sonr::system::System;
use sonr::net::tcp::ReactiveTcpListener;

fn main() -> Result<()> {
    System::init();
    let first_listener = ReactiveTcpListener::bind("127.0.0.1:5000")?;
    let second_listener = ReactiveTcpListener::bind("127.0.0.1:5001")?;
    let server = first_listener.and(second_listener);
    System::start(server);
}

fn map<F, T>(self, callback: F) -> Map<Self, F, T>

Capture the output of a reactor in a closure.

// Create a listener, print the address every time
// the listener accepts a new connection, then push that
// connection onto a queue.
System::init();
let listener = tcp::ReactiveTcpListener::bind("127.0.0.1:5000")?;
let queue = ReactiveQueue::unbounded();
let server = listener.map(|(stream, addr)| {
    stream
}).chain(queue);
System::start(server);

Loading content...