Struct sonr::PreVec

pub struct PreVec<T> { /* fields omitted */ }

PreVec: a collection that allows inserts at a specific index with an optional offsets. A PreVec can be created with either capacity or capacity_with_offset.

This is useful with multiple collections where each index has to be unique, e.g having two collections of connections driven by the same Poll instance where the index represents the connection token.

By default the collection can grow; To prevent overlap with another collection call prevent_growth. This will result in an error if an insert occurrs outside the allocated space.

Example

Creating two PreVecs sharing a unique range:

use sonr::PreVec;

let mut lower = PreVec::with_capacity(1000);
let mut upper = PreVec::with_capacity_and_offset(1000, 1000);

let lower_index = lower.insert("foo").unwrap();
let upper_index = upper.insert("bar").unwrap();

assert_eq!(lower_index, 0);
assert_eq!(upper_index, 1000);

Note that setting the offset to std::usize::MAX would mean only one element can be stored in the PreVec, and any attempt to insert more would result in an error

Methods

impl<T> PreVec<T>

pub fn with_capacity(cap: usize) -> Self

Create a PreVec with a set capacity. Inserting above capacity will either allocate more space or return an error depending on wheter enable_growth or prevent_growth is called,

By default growth is enabled.

Example

let v = PreVec::<u32>::with_capacity(10);
assert_eq!(v.capacity(), 10);

pub fn with_capacity_and_offset(cap: usize, offset: usize) -> Self

Create a PreVec with a set capacity and offset.

Example

let mut v = PreVec::<u32>::with_capacity_and_offset(10, 5);
assert_eq!(v.offset(), 5);

let index = v.insert(5).unwrap();
assert_eq!(index, 5);

pub fn capacity(&self) -> usize

Return the capacity

pub fn offset(&self) -> usize

Return the offset

pub fn prevent_growth(&mut self)

Prevent inserting above the capacity.

pub fn enable_growth(&mut self)

Enable the collection to grow and allocate more space.

pub fn in_range(&self, index: usize) -> bool

Check if the index is within the range of the collection (between capacity and offset)

pub fn set_offset(&mut self, new_offset: usize)

Set the offset of the collection.

pub fn get(&self, index: usize) -> Option<&T>

Get an entry at a specific index

Example

let mut v = PreVec::with_capacity_and_offset(2, 10);
v.insert(1u32);
assert_eq!(v.get(10), Some(&1));

pub fn get_mut(&mut self, index: usize) -> Option<&mut T>

Get a mutable entry at a specific index

Example

let mut v = PreVec::with_capacity_and_offset(2, 10);
v.insert("original".to_string());
v.get_mut(10).map(|s| *s = "changed".to_string()).unwrap();
assert_eq!(&v[10], "changed");

pub fn insert(&mut self, v: T) -> Result<usize>

Insert a value in the next available slot and return the slot index. Inserting above capacity will cause reallocation if and only if the PreVec can grow.

Example

let mut v = PreVec::with_capacity_and_offset(2, 10);
let index = v.insert(1).unwrap();
assert_eq!(index, 10);

pub fn remove(&mut self, index: usize) -> Option<T>

Remove at index (inserting an empty entry)

pub fn len(&self) -> usize

Number of occupied slots

pub fn is_empty(&self) -> bool

Returns true if the collection has no entries

pub fn clear(&mut self)

Remove all entries