Struct sonr::sync::queue::Dequeue

pub struct Dequeue<T> { /* fields omitted */ }

An evented work stealing dequeue

Methods

impl<T> Dequeue<T>

pub fn bounded(stealer: Stealer<T>, capacity: usize) -> Self

Create a bounded dequeue. Setting the capacity to zero means no data will be held in the queue and the publishing thread will blocked until the data is picked up by the dequeue.

pub fn unbounded(stealer: Stealer<T>) -> Self

Create an unbounded dequeue

pub fn sender(&self) -> SignalSender<()>

Get the signal sender that notifies the dequeue of possible new data in the queue.

pub fn steal(&self) -> Steal<T>

Attempt to steal data