Crate sonr

Simple Opinionated Networking in Rust

Sonr builds on the idea of chaining reactors together to form the flow of the application.

TcpListener -> Authentication -> Chat

Modules

errors
SONR default Error

net

prelude

reactor

sync

system

Structs

Event
An readiness event returned by Poll::poll.

Poll
Polls for readiness events on all registered values.

PollOpt
Options supplied when registering an Evented handle with Poll

PreVec
PreVec: a collection that allows inserts at a specific index with an optional offsets. A PreVec can be created with either capacity or capacity_with_offset.

Ready
A set of readiness event kinds

Token
Associates readiness notifications with Evented handles.

Traits

Evented
A value that may be registered with Poll